fix: resolve Windows build and CI smoke test failures
- Change MAX_INSTANCES from u32 to usize for tokio 1.49 compatibility
- Create test.js file before smoke tests in CI workflow

+++ crates/fastnode-daemon/src/ipc/windows.rs
@@ -16,7 +16,7 @@ use tokio::sync::Mutex;
 const PIPE_PREFIX: &str = r"\\.\pipe\";
 
 /// Maximum number of server instances (concurrent connections).
-const MAX_INSTANCES: u32 = 64;
+const MAX_INSTANCES: usize = 64;
 
 /// Connection timeout for clients.
 const CONNECT_TIMEOUT: Duration = Duration::from_secs(2);
